Refactor AxiosInstance and createAxiosInstance functions
