+++ copilot-widget/lib/data/chat.ts
@@ -1,0 +1,25 @@
+import { InitialDataType } from "@lib/types";
+import axios, { AxiosInstance } from "axios";
+
+export function createAxiosInstance(apiUrl?: string, sessionId?: string | null, botToken?: string) {
+    const instance = axios.create({
+        baseURL: apiUrl,
+        headers: {
+            "X-Session-Id": sessionId,
+            "X-Bot-Token": botToken,
+        },
+    });
+
+    instance.interceptors.request.use((config) => {
+        config.data = {
+            ...config.data,
+            session_id: sessionId,
+        };
+        return config;
+    });
+    return instance;
+}
+
+export async function getInitialData(instance: AxiosInstance) {
+    return (await instance.get<InitialDataType>("/chat/init")).data;
+}
